+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * :sparkles: Fragmentx adds requireContext, requireAppContent method
 * :sparkles: Activityx adds appContext method
 * :sparkles: Viewx adds appContext method
+* :fire: Intentx canStartActivity and safeStartActivity method removed to Activityx and rename to canStart and safeStart
 
 
 ## v0.3.2